requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.0
django-extensions==3.2.3
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
whitenoise==6.6.0
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
channels==4.0.0
sentry-sdk==1.39.1